+++ driver/targetmachine.cpp
@@ -57,7 +57,20 @@ static llvm::cl::opt<bool, true> preserveDwarfLineSection(
     llvm::cl::init(false));
 #endif
 
-const char *getABI(const llvm::Triple &triple, const llvm::SmallVector<llvm::StringRef, 8> &features) {
+// Returns true if 'feature' is enabled and false otherwise. Handles the
+// case where the feature is specified multiple times ('+m,-m'), and
+// takes the last occurrence.
+bool isFeatureEnabled(const llvm::SmallVectorImpl<llvm::StringRef> &features,
+                      llvm::StringRef feature) {
+  for (auto it = features.rbegin(), end = features.rend(); it != end; ++it) {
+    if (it->substr(1) == feature) {
+      return (*it)[0] == '+';
+    }
+  }
+  return false;
+};
+
+const char *getABI(const llvm::Triple &triple, const llvm::SmallVectorImpl<llvm::StringRef> &features) {
   llvm::StringRef ABIName(opts::mABI);
   if (ABIName != "") {
     switch (triple.getArch()) {
@@ -111,17 +124,6 @@ const char *getABI(const llvm::Triple &triple, const llvm::SmallVector<llvm::Str
             ABIName.str().c_str());
   }
 
-  // checks if the features include ±<feature>
-  auto hasFeature = [&features](llvm::StringRef feature) {
-    for (int i = features.size() - 1; i >= 0; i--) {
-      auto f = features[i];
-      if (f.substr(1) == feature) {
-	return f[0] == '+';
-      }
-    }
-    return false;
-  };
-
   switch (triple.getArch()) {
   case llvm::Triple::mips64:
   case llvm::Triple::mips64el:
@@ -131,9 +133,9 @@ const char *getABI(const llvm::Triple &triple, const llvm::SmallVector<llvm::Str
   case llvm::Triple::ppc64le:
     return "elfv2";
   case llvm::Triple::riscv64:
-    if (hasFeature("d"))
+    if (isFeatureEnabled(features, "d"))
       return "lp64d";
-    if (hasFeature("f"))
+    if (isFeatureEnabled(features, "f"))
       return "lp64f";
     return "lp64";
   case llvm::Triple::riscv32:
@@ -445,13 +447,9 @@ createTargetMachine(const std::string targetTriple, const std::string arch,
 
   // checks if the features include ±<feature>
   auto hasFeature = [&features](llvm::StringRef feature) {
-    for (int i = features.size() - 1; i >= 0; i--) {
-      auto f = features[i];
-      if (f.substr(1) == feature) {
-	return f[0] == '+';
-      }
-    }
-    return false;
+    return std::any_of(
+        features.begin(), features.end(),
+        [feature](llvm::StringRef f) { return f.substr(1) == feature; });
   };
 
   // cmpxchg16b is not available on old 64bit CPUs. Enable code generation
@@ -462,14 +460,12 @@ createTargetMachine(const std::string targetTriple, const std::string arch,
 
   // For a hosted RISC-V 64-bit target default to rv64gc if nothing has
   // been selected
-  if (triple.getArch() == llvm::Triple::riscv64 &&
-      triple.getOS() != llvm::Triple::UnknownOS &&
-      features.empty()) {
-    features.push_back("+m");
-    features.push_back("+a");
-    features.push_back("+f");
-    features.push_back("+d");
-    features.push_back("+c");
+  if (triple.getArch() == llvm::Triple::riscv64 && features.empty()) {
+    const llvm::StringRef os = triple.getOSName();
+    const bool isFreeStanding = os.empty() || os == "unknown" || os == "none";
+    if (!isFreeStanding) {
+      features = {"+m", "+a", "+f", "+d", "+c"};
+    }
   }
 
   // Handle cases where LLVM picks wrong default relocModel